requirements.txt
requests
python-dotenv
pillow
pysubs2
fonttools